+++ src/enums/webrtc/sctp_transport_state.cc
@@ -1,0 +1,5 @@
+#include "src/enums/webrtc/sctp_transport_state.h"
+
+#define ENUM(X) SCTP_TRANSPORT_STATE ## X
+#include "src/enums/macros/impls.h"
+#undef ENUM